python/requirements.txt
python-telegram-bot==20.7
openai==1.23.2
python-dotenv
requests